Reorganise nAMESPACES AND NAMES.
