+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/AlpacaDevice.cs
