+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/AlpacaDevices/AlpacaCamera.cs
