+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/AlpacaDevices/AlpacaFilterWheel.cs
