+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/AlpacaDevices/AlpacaTelescope.cs
