+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/AlpacaDevices/BaseClass/AlpacaDeviceBaseClass.cs
