+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/AxisRates.cs
