+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/EntityClasses/NoReturnValue.cs
