+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/Enums/ServiceType.cs
