+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/Rate.cs
