+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/Shared Constants.cs
