+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Clients/TrackingRates.cs
