+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/AlpacaDiscovery.cs
