+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/AlpacaDiscoveryResponse.cs
