+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Constants.cs
