+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Entity Classes/AlpacaConfiguredDevice.cs
