+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Entity Classes/AlpacaDevice.cs
