+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Entity Classes/AlpacaDeviceDescription.cs
