+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Entity Classes/AscomDevice.cs
