+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Entity Classes/DnsResponse.cs
