+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Finder.cs
