+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/README.md
