+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Responder.cs
