+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Response Classes/AlpacaConfiguredDevicesResponse.cs
