+++ ASCOM.Standard.AlpacaClients/ASCOM.Alpaca.Discovery/Response Classes/AlpacaDescriptionResponse.cs
