+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/BaseClass/ASCOMDevice.cs
