+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/BaseClass/DynamicAccess.cs
