+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Camera.cs
