+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/CoverCalibrator.cs
