+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Dome.cs
