+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/FilterWheel.cs
