+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Focuser.cs
