+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/ObservingConditions.cs
