+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Profile/ASCOMRegistration.cs
