+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Profile/DriverTypes.cs
