+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Rotator.cs
