+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/SafetyMonitor.cs
