+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Switch.cs
