+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Telescope.cs
