+++ ASCOM.Standard.COM.DriverAccess/ASCOM.Com.DriverAccess/Video.cs
