+++ Directory.Build.props
@@ -2,6 +2,6 @@
  <PropertyGroup>
 	<!-- This is the solution level version for Nuget packages, This should be x.x.x semver (or x.x.x-suffix for pre-release -->
 	<!-- https://docs.microsoft.com/en-us/nuget/concepts/package-versioning -->
-   <Version>1.0.0-alpha21j</Version>
+   <Version>1.0.0-alpha21k</Version>
  </PropertyGroup>
 </Project>
